fix(pii): restore whitespace/country-code prefilter coverage; split module
The byte prefilter dropped three PII cases the old resident RegexSet redacted,
because its candidate gates were narrower than the precise regexes they guard:

- Bare country-code NANP (`12025551234`): old SCREEN reached PHONE_NANP_RE via
  the `\d{11,}` run. The new `phone_nanp` gate keyed only on an internal
  separator, and the bare-CPF fallback's checksum rejects it — so it was left
  unredacted. Gate now also fires on `max_digit_run >= 11` (a bare 10-digit run
  still stays out of the phone path).
- English `My\s?Number` with a tab/newline separator: the keyword needle scan
  only matched a literal-space `my number`. Now gates on both `my` and `number`
  substrings, covering every `\s` variant.
- Newline-separated Aadhaar (`2341\n2341\n2346`): `has_space` tracked only
  space/tab, but the precise `[\s-]` separator class includes `\n`/`\r`/etc.
  Renamed to `has_ws` covering all ASCII whitespace.

Also splits the oversized `pii.rs` (762 -> 475 lines) into focused siblings
`pii/prefilter.rs` and `pii/normalize.rs`, per the <500-line repo guideline.

Adds regression tests for all three cases and extends the
`prefilter_is_superset_of_legacy_screen` corpus with the whitespace variants.

Addresses Codex review comments on PR #119.
