+++ src/memory/store/safety/pii/normalize.rs
@@ -1,0 +1,79 @@
+//! Unicode normalization for PII matching.
+//!
+//! A pre-pass that defeats fullwidth-digit and zero-width-char bypasses while
+//! keeping a byte map back to the original string, so matches found on the
+//! normalized view can be spliced onto the exact original bytes.
+
+pub(super) struct NormalizedView {
+    pub(super) normalized: String,
+    // For each byte offset i in `normalized`, `byte_map[i]` is the byte offset
+    // in the original string where the corresponding char *starts*.
+    // The last entry maps the normalized length to the original length, so
+    // `norm_to_orig(normalized.len())` is well-defined.
+    byte_map: Vec<usize>,
+}
+
+impl NormalizedView {
+    pub(super) fn build(original: &str) -> Self {
+        let mut normalized = String::with_capacity(original.len());
+        let mut byte_map: Vec<usize> = Vec::with_capacity(original.len() + 1);
+        for (idx, ch) in original.char_indices() {
+            if is_zero_width(ch) {
+                continue;
+            }
+            let mapped = fold_char(ch);
+            let start = normalized.len();
+            normalized.push(mapped);
+            // One byte_map entry per byte of the normalized char.
+            let added = normalized.len() - start;
+            for _ in 0..added {
+                byte_map.push(idx);
+            }
+        }
+        byte_map.push(original.len());
+        Self {
+            normalized,
+            byte_map,
+        }
+    }
+
+    pub(super) fn norm_to_orig(&self, norm_byte: usize) -> usize {
+        if norm_byte >= self.byte_map.len() {
+            return *self.byte_map.last().unwrap_or(&0);
+        }
+        self.byte_map[norm_byte]
+    }
+}
+
+fn is_zero_width(c: char) -> bool {
+    matches!(
+        c,
+        '\u{200B}'
+            | '\u{200C}'
+            | '\u{200D}'
+            | '\u{200E}'
+            | '\u{200F}'
+            | '\u{2060}'
+            | '\u{180E}'
+            | '\u{FEFF}'
+    )
+}
+
+fn fold_char(c: char) -> char {
+    match c {
+        // Fullwidth digits 0-9
+        '\u{FF10}'..='\u{FF19}' => char::from_u32(c as u32 - 0xFF10 + 0x30).unwrap_or(c),
+        // Arabic-Indic digits ٠-٩
+        '\u{0660}'..='\u{0669}' => char::from_u32(c as u32 - 0x0660 + 0x30).unwrap_or(c),
+        // Eastern Arabic-Indic digits ۰-۹
+        '\u{06F0}'..='\u{06F9}' => char::from_u32(c as u32 - 0x06F0 + 0x30).unwrap_or(c),
+        // Common fullwidth punctuation we care about for PII formats
+        '\u{FF0D}' => '-',
+        '\u{FF0E}' => '.',
+        '\u{FF0F}' => '/',
+        '\u{FF1A}' => ':',
+        '\u{2010}'..='\u{2015}' => '-', // various unicode hyphens/dashes
+        '\u{2212}' => '-',              // minus sign
+        other => other,
+    }
+}
